# demo (1)/demo/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Open-api
  description: Demo Open_api
  version: 0.0.1
servers:
  - url: http://localhost:9090
    description: Generated server url
paths:
  /api/patient/update/{id}:
    put:
      tags:
        - patient-controller
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/patient/register:
    post:
      tags:
        - patient-controller
      operationId: registerPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/patient/{id}:
    get:
      tags:
        - patient-controller
      operationId: getPatientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/patient/all:
    get:
      tags:
        - patient-controller
      operationId: getAllPatients
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/patient/delete/{id}:
    delete:
      tags:
        - patient-controller
      operationId: deletePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  schemas:
    RegisterRequestDTO:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        gender:
          type: string
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        notificationCount:
          type: object